requirements.txt

django==3.0.9
django-heroku
dj-database-url
django-crispy-forms==1.9.2
django-djconfig==0.10.0
django-haystack==3.0b1
django-infinite-scroll-pagination==1.1.0
django-js-asset==1.2.2
django-machina==1.1.2
django-mptt==0.11.0
django-widget-tweaks==1.4.8
markdown2==2.3.9
mistune==0.8.4
olefile==0.46
Pillow==7.0.0
psycopg2-binary==2.8.5
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.2.0
Whoosh==2.7.4
django-environ==0.4.5
sites_framework
